+++ cmd/influxd/config.go
@@ -110,6 +110,13 @@ type Config struct {
 		RaftTracing  bool   `toml:"raft-tracing"`
 	} `toml:"logging"`
 
+	Statistics struct {
+		Enabled         bool     `toml:"enabled"`
+		Database        string   `toml:"database"`
+		RetentionPolicy string   `toml:"retention-policy"`
+		WriteInterval   Duration `toml:"write-interval"`
+	}
+
 	ContinuousQuery struct {
 		// when continuous queries are run we'll automatically recompute previous intervals
 		// in case lagged data came in. Set to zero if you never have lagged data. We do
@@ -165,6 +172,11 @@ func NewConfig() *Config {
 	c.ContinuousQuery.ComputeNoMoreThan = Duration(2 * time.Minute)
 	c.ContinuousQuery.Disable = false
 	c.ReportingDisabled = false
+
+	c.Statistics.Enabled = false
+	c.Statistics.Database = "_internal"
+	c.Statistics.RetentionPolicy = "default"
+	c.Statistics.WriteInterval = Duration(1 * time.Minute)
 
 	// Detect hostname (or set to localhost).
 	if c.Hostname, _ = os.Hostname(); c.Hostname == "" {